\name{timepoint-package}
\alias{timepoint-package}
\alias{timepoint}
\docType{package}
\title{
  Temporal Event Modeling Using Point Processes
}
\description{
  Offers models and utilities for event time data using point
      processes.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Maximilian Aigner <maximilian.aigner@unil.ch>.

%Maintainer: Your Name <your@email.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
